Se agregaron imports y se agregó la función read_series, en database.py. En main.py se colocaron ifs en lugar de match para cada una de las selecciones.

+++ main.py
@@ -4,11 +4,15 @@
 #   
 #   
 from sre_parse import State
+from turtle import clear
 from IO_methods import input_new_series, input_update_series, menu_builder, clean_screen
 import database
+from database import read_series
+
+
 
 while True:
-    clean_screen()
+    #clean_screen()
     selected_option = menu_builder("NOW, WHAT YOU WANNA DO?",
                                     ["Add new series",
                                     "Delete series",
@@ -18,22 +22,22 @@ while True:
                                     "Move around with <DOWN ARROW>, <UP ARROW>, confirm with <ENTER>",
                                     return_value=True, default_options=["Exit"])
 
-    match selected_option:
-        case "Add new series":
+    if selected_option == "Add new series": 
             new_item_data = input_new_series()
             database.add_series(*new_item_data)
-        case "Delete series":
+    if selected_option == "Delete series":
             selected_series = menu_builder("Which series you want to delete?",
                                             database.get_series_items_names(),
                                             "Move around with <DOWN ARROW>, <UP ARROW>, confirm with <ENTER>",
                                             return_value=True, default_options=["Exit"])
             database.delete_series(selected_series)
-        case "Edit tracked series":
+    if selected_option == "Edit tracked series":
             update_item_data = input_update_series()
             database.update_series(*update_item_data)
-        case "View tracked series":
+    if selected_option == "View tracked series":
+           read_series()
+           #break
+    if selected_option == "Show Statistics":
             pass
-        case "Show Statistics":
-            pass
-        case "Exit":
+    if selected_option == "Exit":
             exit()
